+++ src/components/App/App.js
@@ -2,14 +2,12 @@ import React from "react";
 import "./App.css";
 import { withFirebase } from "utils/Firebase";
 import { compose } from "recompose";
-import ExampleComponent from "ExampleComponent/ExampleComponent";
-import AssignmentPage from "../../pages/AssignmentPage/AssignmentPage";
-import StudentList from "../StudentList/StudentList";
+import CreateAssignment from "../../pages/CreateAssignment/CreateAssignment";
 
 function App() {
   return (
     <div className="App">
-      <AssignmentPage />
+      <CreateAssignment />
     </div>
   );
 }
