+++ src/components/WordGroupIcon/WordGroupIcon.scss
